+++ src/lib/Config.ts
@@ -1,9 +1,5 @@
-import type { ThemeConfig } from '$lib';
-import { carbon } from '$lib/themes';
-
 /** */
 export type AdminConfig = {
-	theme: ThemeConfig;
 	defaultLocale: string;
 	autoCloseSideMenu: boolean;
 	rootUrl: string;
@@ -15,7 +11,6 @@ export type AdminConfig = {
 
 export function defaultAdminConfig(): AdminConfig {
 	return {
-		theme: carbon,
 		defaultLocale: 'en',
 		autoCloseSideMenu: false,
 		rootUrl: '/',