+++ src/lib/Fields/Email.ts
@@ -2,8 +2,7 @@ import type { InputFieldOptions, FormFieldTheme, ViewFieldTheme } from '$lib';
 import { BaseField } from '$lib/Fields';
 
 /** */
-export type EmailFieldOptions = InputFieldOptions & {
-};
+export type EmailFieldOptions = InputFieldOptions & {};
 
 /** */
 export class EmailField extends BaseField<EmailFieldOptions> {